add search functionality

+++ Commands/RemoveTransactionCommand.cs
@@ -8,16 +8,22 @@ public class RemoveTransactionCommand : Command {
         List<Transaction> transactions = transactionService.GetAllTransactions();
 
         int i = 0;
-        for (i = 0; i < transactions.Count; i++ ) {
-            Console.WriteLine($"TRANSACTION\nindex: {i}\n{transactions[i].ToString()}");
+        for (i = 0; i < transactions.Count; i++) {
+            Console.WriteLine($"TRANSACTION\nindex: {i}\n{transactions[i]}");
             Console.WriteLine("");
         }
 
-        Console.WriteLine("Enter index of transaction to remove:");
+        Console.WriteLine("Enter the index of the transaction to remove:");
+        string? userInput = Console.ReadLine();
 
-        int input = int.Parse(Console.ReadLine());
+        if (int.TryParse(userInput, out int input) && input >= 0 && input < transactions.Count) {
 
-        transactionService.RemoveTransaction(transactions[input].Id);
-        Console.WriteLine("Removed transaction succesfully.");
-    }
+            transactionService.RemoveTransaction(transactions[input].Id);
+            Console.WriteLine("Removed transaction successfully.");
+        }
+        else {
+
+            Console.WriteLine("Invalid input. Please enter a valid transaction index.");
+        }
+    }      
 }
